% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PioneerMetastaticAE.R
\docType{package}
\name{PioneerMetastaticAE-package}
\alias{PioneerMetastaticAE}
\alias{PioneerMetastaticAE-package}
\title{PioneerMetastaticAE: A Study Package for Generating Cohort Diagnostics}
\description{
A package that generates cohort diagnostics for cohort definitions included in the package.
}
\keyword{internal}
